resolving PR comment
